+++ extensions/src/extensions.ts
@@ -4,3 +4,4 @@ export * from './extensions/reinitializeOptions';
 export * from './extensions/setReadonly';
 export * from './extensions/setValue';
 export * from './extensions/valueChange';
+export * from './extensions/patchOptions';