+++ extensions/src/extensions/reinitializeOptions.ts
@@ -4,7 +4,7 @@ import {NgSelectAction, NgSelectOptions} from "@anglr/select";
  * Method reinitialize options for NgSelect
  * @param options - Options to be used for reinitialization
  */
-export function reinitializeOptions<TValue>(options?: NgSelectOptions<TValue>): NgSelectAction<TValue>
+export function reinitializeOptions<TValue = any>(options?: NgSelectOptions<TValue>): NgSelectAction<TValue>
 {
     return ngSelect =>
     {