+++ src/misc/pluginBus/pluginBus.ts
@@ -1,12 +1,14 @@
-import {Injectable, EventEmitter} from '@angular/core';
+import {Injectable, EventEmitter, ElementRef} from '@angular/core';
 
 import {PluginBusOptions} from '../ngSelectOptions.interface';
+import {NgSelectOption} from '../../components/option';
+import {PluginBusEvents} from './pluginBus.interface';
 
 /**
  * Class represents centralized bus for plugin communications
  */
 @Injectable()
-export class PluginBus<TValue = any>
+export class PluginBus<TValue = any> implements PluginBusEvents
 {
     //######################### public properties #########################
 
@@ -15,22 +17,40 @@ export class PluginBus<TValue = any>
      */
     public selectOptions: PluginBusOptions<TValue>;
 
+    /**
+     * HTML element that represents select itself
+     */
+    public selectElement: ElementRef<HTMLElement>;
+
     //######################### public properties - events #########################
 
     /**
      * Occurs when popup visibility should be toggled
      */
-    public togglePopup: EventEmitter<void>;
+    public togglePopup: EventEmitter<void> = new EventEmitter<void>();
 
     /**
      * Occurs when popup visibility should be changed
      */
-    public showHidePopup: EventEmitter<boolean>;
+    public showHidePopup: EventEmitter<boolean> = new EventEmitter<boolean>();
 
     /**
      * Occurs when popup visibility should be changed
      */
-    public popupVisibilityChange: EventEmitter<void>;
+    public popupVisibilityChange: EventEmitter<void> = new EventEmitter<void>();
 
-    //######################### public properties - state #########################
+    /**
+     * Occurs when option should be selected
+     */
+    public optionSelect: EventEmitter<NgSelectOption> = new EventEmitter<NgSelectOption>();
+
+    /**
+     * Occurs when option should be canceled
+     */
+    public optionCancel: EventEmitter<NgSelectOption> = new EventEmitter<NgSelectOption>();
+
+    /**
+     * Occurs when any part of select gains focus
+     */
+    public focus: EventEmitter<void> = new EventEmitter<void>();
 }