+++ src/select.ts
@@ -11,6 +11,7 @@ export * from './misc/ngSelectControlValueAccessor.directive';
 export * from './misc';
 export * from './misc/types';
 export * from './misc/pluginBus/pluginBus';
+export * from './misc/pluginBus/pluginBus.interface';
 export * from './misc/optionsGatherer/types';
 export * from './modules/ngSelect.module';
 export * from './modules/ngSelectDynamic.module';