+++ changelog.md
@@ -3,8 +3,14 @@
 ## Version 8.0.0
 
 - all generic types are now optional, all defaults to `any`
+- added new method `listenTo` to `NgSelect` allowing subscribing to select events
+- added new method `initializePluginInstance` to `NgSelect` allowing initialization of external plugin instance
+- added `PluginBus` which serves serves as bus for shared events also shared select options and gatherers
+- added new extension method `patchOptions` which allows changing options without running initialization
+- added `placeholderNothingSelected` into `texts` for `EditLiveSearch` options
 - **Breaking Changes**
-   - test
+   - all plugins now have to have `pluginBus`
+   - `KeyboardHandler` handler removed `selectElement`, `optionsGatherer`, `popupVisibilityRequest`, `optionSelect` properties, use new `PluginBus` instead
 
 ## Version 7.2.0